@@ -18,28 +18,28 @@ allprojects {
 }
 
 ext {
-    minSdkVersion = 11
-    compileSdkVersion = 27
+    minSdkVersion = 14
+    compileSdkVersion = 29
     targetSdkVersion = compileSdkVersion
 
-    rxJava = 'io.reactivex.rxjava2:rxjava:2.1.16'
-    supportLibraryVersion = '27.1.1'
-    appCompat = "com.android.support:appcompat-v7:$supportLibraryVersion"
-    supportAnnotations = "com.android.support:support-annotations:$supportLibraryVersion"
-    supportFragment = "com.android.support:support-fragment:$supportLibraryVersion"
-    junit = 'junit:junit:4.12'
-    mockito = 'org.mockito:mockito-core:1.10.19'
+    rxJava = 'io.reactivex.rxjava3:rxjava:3.0.4'
+    junit = 'junit:junit:4.13'
+    mockito = 'org.mockito:mockito-core:3.3.3'
 
-    robolectricVersion = '3.3.2'
+    androidXFragment = 'androidx.fragment:fragment:1.2.5'
+    androidXAnnotation = 'androidx.annotation:annotation:1.1.0'
+    androidXAppcompat =  'androidx.appcompat:appcompat:1.1.0'
+
+    robolectricVersion = '4.3.1'
     robolectric = "org.robolectric:robolectric:$robolectricVersion"
     robolectricShadowsSupport = "org.robolectric:shadows-support-v4:$robolectricVersion"
     // Workaround for https://github.com/robolectric/robolectric/issues/1932
     khronosOpenGLApi = "org.khronos:opengl-api:gl1.1-android-2.1_r1"
 
     bintrayRepo = 'tbruyelle'
-    bintrayName = 'RxPermissions2'
+    bintrayName = 'RxPermissions3'
 
-    publishedGroupId = 'com.tbruyelle.rxpermissions2'
+    publishedGroupId = 'com.tbruyelle.rxpermissions3'
     artifact = 'rxpermissions'
     libraryName = 'RxPermissions'
     libraryVersion = '0.9.5'
